created classifier plot
