+++ src/utils/plotSpecs/generateMitochondrialFractionHistogram.js
@@ -133,8 +133,8 @@ const generateSpec = (config, plotData) => {
         title: { value: config.axes.yAxisText },
         titleFont: { value: config.fontStyle.font },
         labelFont: { value: config.fontStyle.font },
-        titleFontSize: { value: config.axisTitlesize },
-        labelFontSize: { value: config.axisTicks },
+        titleFontSize: { value: config.axes.titleFontSize },
+        labelFontSize: { value: config.axes.labelFontSize },
         offset: { value: config.axes.offset },
         gridOpacity: { value: (config.axes.gridOpacity / 20) },
       },
